backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Authlib==1.2.1
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
Werkzeug==2.3.7
requests==2.31.0